refactor: merge all routes
