+++ routes/__init__.py
@@ -20,7 +20,6 @@ from .core import (
     upload,
     user_routes,
 )
-from all_routes import routes_bp
 
 # Type variable for blueprints
 BlueprintT = TypeVar('BlueprintT', bound=Blueprint)
@@ -28,31 +27,27 @@ BlueprintT = TypeVar('BlueprintT', bound=Blueprint)
 __all__ = ['ALL_BLUEPRINTS', 'all_blueprints']
 
 # Blueprints must be registered in order of most specific to least specific
-# to prevent route shadowing. The routes_bp contains core API endpoints that
-# should take precedence over other blueprints.
+# to prevent route shadowing. Core routes should come first.
 ALL_BLUEPRINTS: List[BlueprintT] = [
-    # 1. Core API routes (most specific)
-    routes_bp,
-    
-    # 3. User management routes
+    # 1. User management routes
     user_routes,
-    
-    # 4. Authentication routes
+
+    # 2. Authentication routes
     auth,
-    
-    # 5. File upload handling
+
+    # 3. File upload handling
     upload,
-    
-    # 6. General API endpoints
+
+    # 4. General API endpoints
     api,
-    
-    # 7. Diagram-specific routes
+
+    # 5. Diagram-specific routes
     diagrams,
-    
-    # 8. Main application routes
+
+    # 6. Main application routes
     main,
-    
-    # 9. Collaboration features (least specific)
+
+    # 7. Collaboration features (least specific)
     collab,
 ]
